feat(workflows): V2-008 durable workflow studio (definitions, runs, executor, recovery, SSE, canvas UI)
- workflow definitions CRUD with graph validation (six node kinds, explicit edges, cycle rejection) and immutable revisions
- durable runs from pinned definition revision; node states with lease/checkpoint/budget; control actions with Idempotency-Key replay
- node executor (application layer) + Celery task/beat wiring, maintenance endpoint and lifespan recovery hook
- budget reserve/record/release with BUDGET_BLOCKED pause; no half-committed versions
- SSE replay (Last-Event-ID) to live tail; frontend snapshot-then-resume wiring, React Flow canvas payload mapping and status stamps
- v1 workflow controls gain Idempotency-Key per blueprint
- export manifest model/repository registrations ride along (executor export node persists manifests; export API surface lands in the V2-009 commit)
- package.json/pnpm-lock add @xyflow/react here; i18next/react-i18next are consumed by the V2-009 commit

+++ proseforge/infrastructure/database/migrations/versions/0013_workflow_definitions.py
@@ -1,7 +1,7 @@
 """Add versioned workflow definitions and per-node run state."""
 
 from alembic import op
-from sqlalchemy import Column, DateTime, Integer, String, Text, UniqueConstraint, inspect
+from sqlalchemy import Column, DateTime, Float, Integer, String, Text, UniqueConstraint, inspect
 
 
 revision = "0013_workflow_definitions"
@@ -38,8 +38,8 @@ def upgrade() -> None:
             Column("retry_count", Integer, nullable=False, server_default="0"),
             Column("reserved_tokens", Integer, nullable=False, server_default="0"),
             Column("used_tokens", Integer, nullable=False, server_default="0"),
-            Column("reserved_cost", Integer, nullable=False, server_default="0"),
-            Column("used_cost", Integer, nullable=False, server_default="0"),
+            Column("reserved_cost", Float, nullable=False, server_default="0"),
+            Column("used_cost", Float, nullable=False, server_default="0"),
             Column("updated_at", DateTime(timezone=True), nullable=False),
             UniqueConstraint("run_id", "node_key", name="uq_workflow_node_states_run_node_key"),
         )
